requirements.txt
requests==2.31.0
python-nmap==0.7.1
pyfiglet==0.8.post1
tabulate==0.9.0
colorama==0.4.6
beautifulsoup4==4.12.2
